mitsu70x:  The reimplemented library now generates sane output!
The backend is no longer WIP.

The Kodak 305 is now fully supported, and the other members of the
family are now considered expected to work (pending testing and
VID/PIDs)

Next up will be to whip the library into distributable shape.  Once
that's set, I'll push the new rendering code into Gutenprint.