shinkos6145:  Support job combining (4x6 *2 -> 8x6-div2)
This code could be genericized, but as the other 6" sinfonia models
automatically combine prints it's not terribly useful.

The 6" printers don't appear to support a 5x7-div2 cut size, so
this remains only useful for 4x6.  Similarly, we can't do a 6x8-div4.

But as all 6" models can rewind, this is only a print speed optimization
rather than a media use optimization!